Straightened out the logic with the node-masks.
Summary: In osg, we use the node-masks for more things than in Coin. The logic, as I see it is that if the object is turned off, no traversals are enabled. If the object is turned on, then the selected traversals are enabled.

Test Plan: Tested on mac. Seems to work.

Reviewers: jaap

Reviewed By: jaap

Differential Revision: https://phabricator.dgbes.com:8443/D358

git-svn-id: https://svn.opendtect.org/od/branches/od_osg@29117 a985855d-d384-4b45-ae33-31a410998c3c