Renamed functions and variables on visSurvey::Scene, with regards to zstretch vis zscale.
Summary:
I wanted to make it clear what is temporary z-stretch changes and permanent ones, so variable names
and function names were changed to make it clear.

Test Plan: Visual check only, plus comments on if it clearer, or if other names should be changed as well.

Reviewers: ranojay

Reviewed By: ranojay

Maniphest Tasks: T6

Differential Revision: https://phabricator.dgbes.com:8443/D39

git-svn-id: https://svn.opendtect.org/od/branches/od_osg@29202 a985855d-d384-4b45-ae33-31a410998c3c