MarkerSet Issues in visfaultsticksetdisplay
Summary:
Replaced visMarker by visMarkerSet. Please check in the function:isSelectableMarkerInPolySel, do we need do display transform or not.
In the class construction, I don't understand why Ranojay removed the marker creation parts in the revision 27364. I think we need this parts. 
So I recovered it and changed it to visMarkerSet.

Test Plan: Compiles and Tests

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D607

git-svn-id: https://svn.opendtect.org/od/branches/od_osg@29580 a985855d-d384-4b45-ae33-31a410998c3c