Bugfix: one-item Text2 object became wrong victim of small feature culling
Summary:
(1) OSG suffers from a chicken-and-egg problem that occurs when computing
the initial bounding box of a screen-sized object (see e.g. code comments
at osgText::TextBase::computeBound()). The initial underestimation of its
scale can make a one-item Text2 object become a victim of small feature
culling for as long as the user does not extremely zoom in on it. Chosen
solution is an implementation of one makeshift suggested by Robert Osfield
on the OSG forum.
(2) Fixed obvious code order error in Text::setDisplayTransformation(.)

Test Plan:
(1) Tested with line name of 2D Seismic line (to be committed). 
(2) Not tested.

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D675

git-svn-id: https://svn.opendtect.org/od/branches/od_osg@29657 a985855d-d384-4b45-ae33-31a410998c3c