Bugfix: got rid of "undef-color" artifact on edge between seismics and border.
Summary:
Displaying yet old texture on already renewed geometry in texture update stage
has unveiled that border color of the combined data/undef layer was not set.
Because its luminance value was coincidentally matching the default, the border
color itself looked fine, but its flipped alpha value was responsible for the
characteristic "undef-color" artifacts on edge between seismics and border.

Test Plan:
Resize an inline from smaller to bigger, while putting a breakpoint in the
middle of the texture update stage. The "Undef-color" artifact on edge between
seismics and border is replaced by a clear transition.

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D1219

git-svn-id: https://svn.opendtect.org/od/branches/od_osg@30337 a985855d-d384-4b45-ae33-31a410998c3c