Fixed indexing problem found during removal with polygon tool.
Summary: Used the newly made PolygonSlection tool to remove selected parts of the 2D horizons. The indices were not properly calculated during removal. So fixed the void VertexShape::removeAllPrimitiveSets() function, and using it for removing everything and re-creating each time.

Test Plan: Tested in od_osg, removal of any portion of the 2D horizon with the polygon selection now works.

Reviewers: ding

Reviewed By: ding

CC: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D1653

git-svn-id: https://svn.opendtect.org/od/branches/od_osg@31165 a985855d-d384-4b45-ae33-31a410998c3c