Added mouse picking on points and lines. Only triangles were supported so far.
Summary:
Added PolytopeIntersector to enable mouse picking on points and lines. The      
LineSegmentIntersector keeps on doing picking on triangles for performance      
reasons. This required fusing the results of both intersectors, finding out
which result is in front of the other.

Test Plan:
Tested on scene with inlines and well logs. The latter contains non-3D          
polylines. Zooming and rotating and moving the mouse over scene objects,        
seeing in the status bar that information of the correct object is shown,       
also if it is occluding another object. Did not explicitly test for points,     
but these are no different than line endings when being returned by the         
PolytopeIntersector.

Reviewers: ding, kristofer

Reviewed By: ding

Maniphest Tasks: T73

Differential Revision: https://phabricator.dgbes.com:8443/D1646

git-svn-id: https://svn.opendtect.org/od/branches/od_osg@31238 a985855d-d384-4b45-ae33-31a410998c3c