Added functionality for arbitrarily oriented texture rectangle
Summary:
Added functionality for creating texture rectangle with arbitrary orientation,  
where it had to be perpenpicular to a local coordinate frame axis so far. The   
presence of a display transformation makes this far from trivial. Defining      
rotation in terms of vectors (instead of angles or quaternions) at the front    
side of the display transformation is the best way to make it both sound and    
understandable.

Test Plan:
Tested by displaying prestack data in a 3D scene flat-viewer, which is using    
this class. The texture rectangles are now nicely displayed perpendicular to    
both an inline, a crossline, and a few differently oriented 2D lines.

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D1939

git-svn-id: https://svn.opendtect.org/od/branches/od_osg@31518 a985855d-d384-4b45-ae33-31a410998c3c