Right mouse click on seismics did not popup menu when dragger active.
Summary:
Right mouse click in order to popup a menu on a seismics section/volume was
stolen by its active osg-dragger.

Test Plan:
Tested on inline/crossline/timeslices and prestack flatviewer slices. 
Dragger is still working, and right-click menu now pops up if dragger
is active as well. Volumes not yet ready to be tested on this.

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D1941

git-svn-id: https://svn.opendtect.org/od/branches/od_osg@31519 a985855d-d384-4b45-ae33-31a410998c3c