Bugfixes for nearly-black texture parts on horizons when shading turned off.
Summary:
(1) If the mAddTriangleIndexes macro can add the triangle to the current        
strip, by error the triangle is added a second time as a whole. Apparently,     
this has no influence on what is displayed, but it must have influenced the     
computational effort.                                                           
(2) The triangle indexes were not added to the triangle strip in a consistent   
order (clockwise for the odd triangles and anticlockwise for the even triangles 
or the other way around), especially because in some border cases not two but   
only one triangle is added per grid patch. This caused a sign flip of the local 
normal directions up to the next inconsistency. Texture mapping with shaders    
apparently does not care about this sign, but texture mapping without shaders   
does.

Test Plan:
Tested both with shading on and off for several F3 demo horizons at different   
resolutions. The barcode-like nearly-black texture overlay patterns are gone.

Reviewers: ding

Reviewed By: ding

Differential Revision: https://phabricator.dgbes.com:8443/D2074

git-svn-id: https://svn.opendtect.org/od/branches/od_osg@31661 a985855d-d384-4b45-ae33-31a410998c3c