od6.0: Fixed wrong cast
Summary:
chunksize_ is set to 0x800000000. When casting to an unsigned int, the value will be 0. This was the reason why preloading volumes larger than 16GB as 32bit, did not work.

I don't see why we would cast to an unsigned int, so changed it to od_uint64

Test Plan:
Could not test data preloading as I only have 16GB RAM in my pc.
While debugging I changed the variable sz to a very large number. allocsize is good after the condition I changed.

Reviewers: raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com:8443/D9158