od6.2: Ignore double-clicks on checkboxes
Summary:
Double-clicks on treeitems pop-up property windows in some cases. A double-click on the treeitem's checkbox should not do that. Qt does not have tools to say whether the click was on the checkbox or on the text itself. The QMouseEvent can give me a relative position of where the user clicked, so used that.
Should fix OD-1789

Test Plan: Added 2D horizon to tree (3D scene and basemap). Double (or quick) click on checkbox does not popup property window. A double-click on text does.

Reviewers: raman, arnaud

Reviewed By: arnaud

Differential Revision: https://phabricator.dgbes.com:8443/D12147