od6.2: fixed inherent limitations to BasicInterval::limitTo
Summary: the base function is not aware of the step. Thus if you try to limit a StepInterval [0.000 5.000] step 0.004 to the StepInterval [0.006 4.994] step 0.004, you will get [0.006 4.994], which results in shifting all samples by 2ms. That is not the goal of a call to limitTo. If you agree I would make the function virtual in the trunk; Here in od6.2 I can only adapt the usage in VolProc::Step.

Test Plan: tested PCA VolProc in F10 survey. Works now.

Reviewers: bert

Reviewed By: bert

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com:8443/D12372