[od6.2] Wrong unit check when creating well logs
Summary: The unit check was done on the output unit, not on the formula output unit. This prevented saving logs in non-si units.

Test Plan: Created a Vp log from sonic, where sonic is in microseconds/ft and the output Vp in ft/s. Got an error message before the change, now creation is successfull and result looks OK.

Reviewers: arnaud

Reviewed By: arnaud

Differential Revision: https://phabricator.dgbes.com:8443/D12506