od6.2: reworked interface to display fluid replacement label in uiStratLayerModelDisp
Summary:
deprecated a number of functions and their class members; They are replaced by a new public function with gets its input from arguments. This function should only be called in two cases:
1- A new model is opened.
2- The fluid replacement button it checked/unchecked.

Also fixed the bug that the text annotation did not move when re-sizing the window in the model viewer, and was displayed cropped in the upper right corner of the synthetic viewer. Fixes OD-400.

Test Plan: the correct FR label is now printed in the model and synthetic viewers.

Reviewers: nanne

Reviewed By: nanne

Subscribers: raman, bert

Differential Revision: https://phabricator.dgbes.com:8443/D12522