od6.2: Fixed crash in datapack pre-loading with non-default Z step
Summary: z range snapping is meant to be equivalent to Interval::limitTo but with two objects with a different step. Previous code would change zstart a large, resulting in out of bounds array pointer. Added a lot of pointer checks in debug mode to ensure it is catachable easily.

Test Plan: pre-loaded a volume of SR=4ms at an output step SR=12ms. No more crash.

Reviewers: raman

Reviewed By: raman

Differential Revision: https://phabricator.dgbes.com:8443/D12536